#24
- examples/server: доработки и упрощения
