+++ examples/server/src/apps/desktop.tsx
@@ -1,11 +1,17 @@
+import React from 'react';
 import type { PageAssets } from '@sima-land/isomorph/http-server/types';
 import { createApplication, Resolve } from '@sima-land/isomorph/di';
 import { PresetResponse } from '@sima-land/isomorph/preset/node/response';
 import { KnownToken } from '@sima-land/isomorph/tokens';
 import { Token } from '../tokens';
 import { Api, createApi } from '../services/api';
-import { prepareDesktopPage } from '../pages/desktop';
 import { sauce } from '@sima-land/isomorph/http-client/sauce';
+import { Provider } from 'react-redux';
+import { GlobalDataScript, SsrBridge } from '@sima-land/isomorph/utils/ssr';
+import { DesktopApp as Desktop } from '../components/desktop';
+import { reducer } from '../reducers/app';
+import { configureStore } from '@reduxjs/toolkit';
+import { rootSaga } from '../sagas';
 
 export function DesktopApp() {
   const app = createApplication();
@@ -36,8 +42,37 @@ function providePrepare(resolve: Resolve): () => Promise<JSX.Element> {
   const config = resolve(KnownToken.Config.base);
   const api = resolve(Token.Response.api);
   const sagaMiddleware = resolve(KnownToken.sagaMiddleware);
+  const responseBuilder = resolve(KnownToken.Response.builder);
 
-  return () => prepareDesktopPage({ api, config, sagaMiddleware });
+  return async function prepare() {
+    const store = configureStore({
+      reducer,
+      middleware: [sagaMiddleware],
+    });
+
+    await sagaMiddleware.timeout(3000).run(rootSaga, { api });
+
+    const bridge = SsrBridge.prepare(config.appName);
+
+    // устанавливаем meta в ответ
+    responseBuilder.meta(
+      JSON.stringify({
+        userId: store.getState().user?.id,
+      }),
+    );
+
+    return (
+      <Provider store={store}>
+        <div id={bridge.rootElementId}>
+          <Desktop />
+        </div>
+        <GlobalDataScript
+          property={bridge.serverDataKey}
+          value={{ INITIAL_STATE: store.getState() }}
+        />
+      </Provider>
+    );
+  };
 }
 
 function provideAssets(resolve: Resolve): PageAssets {
