--- examples/server/src/pages/mobile.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export function prepareMobilePage() {
-  return (
-    <div>
-      <h1>Example app</h1>
-      <h2>Mobile version</h2>
-      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Ratione, minima.</p>
-    </div>
-  );
-}
